# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FocusQuest API - ADHD gamified task tracker
servers:
  - url: /api
    description: Base API path
...
paths:
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          description: Relative path to redirect to after login (must start with `/`). Defaults to `/`.
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        '302':
          description: Redirect to the `returnTo` path on success (defaults to `/`) or `/api/login` on failure.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Token did not contain usable claims.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Token exchange failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'
...
  /accountability/partners:
    get:
      operationId: getPartners
      tags: [accountability]
      summary: List accountability partners
      responses:
        "200":
          description: Partners list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Partnership"
    post:
      operationId: sendPartnerRequest
      tags: [accountability]
      summary: Send an accountability partner request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartnerRequestInput"
      responses:
        "201":
          description: Request sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Partnership"
  /accountability/partners/{id}/accept:
    post:
      operationId: acceptPartnerRequest
      tags: [accountability]
      summary: Accept a partnership request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Partnership accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Partnership"
  /accountability/partners/{id}/decline:
    post:
      operationId: declinePartnerRequest
      tags: [accountability]
      summary: Decline or remove a partnership
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Partnership declined/removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Partnership"
  /accountability/partners/{id}/feed:
    get:
      operationId: getPartnerFeed
      tags: [accountability]
      summary: Get recent activity for a partner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Partner activity feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
...
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
          format: uri
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1
    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    UserSummary:
      type: object
      required: [id, username, currentLevel, totalPoints]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: ["string", "null"]
        avatarColor:
          type: string
        currentLevel:
          type: integer
        levelName:
          type: string
        totalPoints:
          type: integer
        streakDays:
          type: integer
...
    Partnership:
      type: object
      required: [id, requesterId, recipientId, status, createdAt]
      properties:
        id:
          type: integer
        requesterId:
          type: integer
        recipientId:
          type: integer
        status:
          type: string
          enum: [pending, accepted, declined]
        partner:
          $ref: "#/components/schemas/UserSummary"
        createdAt:
          type: string
    PartnerRequestInput:
      type: object
      required: [recipientId]
      properties:
        recipientId:
          type: integer
    ActivityItem:
      type: object
      required: [id, userId, type, description, points, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        username:
          type: string
        type:
          type: string
          enum: [task_completed, badge_earned, level_up, streak_milestone, all_day_bonus, streak_freeze_bought, streak_freeze_used]
        description:
          type: string
        points:
          type: integer
        createdAt:
          type: string
...