# lib/api-zod/src/generated/types/errorEnvelope.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FocusQuest API - ADHD gamified task tracker
 * OpenAPI spec version: 0.1.0
 */

export interface ErrorEnvelope {
  error: string;
}